+++ src/EntityList.hpp
@@ -21,7 +21,7 @@ public:
         return _objects;
     }
 
-    virtual bool hit(const Ray &r, double t_min, double t_max, HitRecord &rec) const override;
+    virtual bool hit(const Ray &r, float t_min, float t_max, HitRecord &rec) const override;
 
 private:
     std::vector<std::shared_ptr<Entity>> _objects;
