+++ src/entities/Sphere.hpp
@@ -13,7 +13,7 @@ public:
     Sphere();
     Sphere(Vector3 c, float r, std::shared_ptr<Material> m);
 
-    virtual bool hit(const Ray &r, double t_min, double t_max, HitRecord &rec) const override;
+    virtual bool hit(const Ray &r, float t_min, float t_max, HitRecord &rec) const override;
 
 private:
     Vector3 _center;
